requirements.txt
apache-airflow==2.1.3
beautifulsoup4
pandas
